+++ resources/lib/playqueue/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+Monitors the Kodi playqueue and adjusts the Plex playqueue accordingly
+"""
+from __future__ import absolute_import, division, unicode_literals
+
+from .common import PlaylistItem, PlaylistItemDummy, PlaylistError, PLAYQUEUES
+from .playqueue import PlayQueue
+from .monitor import PlayqueueMonitor
+from .functions import init_playqueues, get_playqueue_from_type, \
+    playqueue_from_plextype, playqueue_from_id, get_PMS_playlist, \
+    init_playqueue_from_plex_children, get_pms_playqueue, \
+    get_plextype_from_xml
