+++ resources/lib/playqueue/monitor.py
@@ -7,113 +7,11 @@ from __future__ import absolute_import, division, unicode_literals
 from logging import getLogger
 import copy
 
-import xbmc
-
-from .plex_api import API
-from . import playlist_func as PL, plex_functions as PF
-from . import backgroundthread, utils, json_rpc as js, app, variables as v
-
-###############################################################################
-LOG = getLogger('PLEX.playqueue')
-
-PLUGIN = 'plugin://%s' % v.ADDON_ID
-
-# Our PKC playqueues (3 instances PlayQueue())
-PLAYQUEUES = []
-###############################################################################
+from .common import PlaylistError, PlaylistItem, PLAYQUEUES
+from .. import backgroundthread, json_rpc as js, utils, app
 
 
-def init_playqueues():
-    """
-    Call this once on startup to initialize the PKC playqueue objects in
-    the list PLAYQUEUES
-    """
-    if PLAYQUEUES:
-        LOG.debug('Playqueues have already been initialized')
-        return
-    # Initialize Kodi playqueues
-    with app.APP.lock_playqueues:
-        for i in (0, 1, 2):
-            # Just in case the Kodi response is not sorted correctly
-            for queue in js.get_playlists():
-                if queue['playlistid'] != i:
-                    continue
-                playqueue = PL.PlayQueue()
-                playqueue.playlistid = i
-                playqueue.type = queue['type']
-                # Initialize each Kodi playlist
-                if playqueue.type == v.KODI_TYPE_AUDIO:
-                    playqueue.kodi_pl = xbmc.PlayList(xbmc.PLAYLIST_MUSIC)
-                elif playqueue.type == v.KODI_TYPE_VIDEO:
-                    playqueue.kodi_pl = xbmc.PlayList(xbmc.PLAYLIST_VIDEO)
-                else:
-                    # Currently, only video or audio playqueues available
-                    playqueue.kodi_pl = xbmc.PlayList(xbmc.PLAYLIST_VIDEO)
-                    # Overwrite 'picture' with 'photo'
-                    playqueue.type = v.KODI_TYPE_PHOTO
-                PLAYQUEUES.append(playqueue)
-    LOG.debug('Initialized the Kodi playqueues: %s', PLAYQUEUES)
-
-
-def get_playqueue_from_type(kodi_playlist_type):
-    """
-    Returns the playqueue according to the kodi_playlist_type ('video',
-    'audio', 'picture') passed in
-    """
-    for playqueue in PLAYQUEUES:
-        if playqueue.type == kodi_playlist_type:
-            break
-    else:
-        raise ValueError('Wrong playlist type passed in: %s'
-                         % kodi_playlist_type)
-    return playqueue
-
-
-def playqueue_from_plextype(plex_type):
-    if plex_type in v.PLEX_VIDEOTYPES:
-        plex_type = v.PLEX_TYPE_VIDEO_PLAYLIST
-    elif plex_type in v.PLEX_AUDIOTYPES:
-        plex_type = v.PLEX_TYPE_AUDIO_PLAYLIST
-    else:
-        plex_type = v.PLEX_TYPE_VIDEO_PLAYLIST
-    for playqueue in PLAYQUEUES:
-        if playqueue.type == plex_type:
-            break
-    return playqueue
-
-
-def playqueue_from_id(kodi_playlist_id):
-    for playqueue in PLAYQUEUES:
-        if playqueue.playlistid == kodi_playlist_id:
-            break
-    else:
-        raise ValueError('Wrong playlist id passed in: %s of type %s'
-                         % (kodi_playlist_id, type(kodi_playlist_id)))
-    return playqueue
-
-
-def init_playqueue_from_plex_children(plex_id, transient_token=None):
-    """
-    Init a new playqueue e.g. from an album. Alexa does this
-
-    Returns the playqueue
-    """
-    xml = PF.GetAllPlexChildren(plex_id)
-    try:
-        xml[0].attrib
-    except (TypeError, IndexError, AttributeError):
-        LOG.error('Could not download the PMS xml for %s', plex_id)
-        return
-    playqueue = get_playqueue_from_type(
-        v.KODI_PLAYLIST_TYPE_FROM_PLEX_TYPE[xml[0].attrib['type']])
-    playqueue.clear()
-    for i, child in enumerate(xml):
-        api = API(child)
-        PL.add_item_to_playlist(playqueue, i, plex_id=api.plex_id())
-    playqueue.plex_transient_token = transient_token
-    LOG.debug('Firing up Kodi player')
-    app.APP.player.play(playqueue.kodi_pl, None, False, 0)
-    return playqueue
+LOG = getLogger('PLEX.playqueue_monitor')
 
 
 class PlayqueueMonitor(backgroundthread.KillableThread):
@@ -169,26 +67,24 @@ class PlayqueueMonitor(backgroundthread.KillableThread):
                     LOG.debug('Playqueue item %s moved to position %s',
                               i + j, i)
                     try:
-                        PL.move_playlist_item(playqueue, i + j, i)
-                    except PL.PlaylistError:
+                        playqueue.plex_move_item(i + j, i)
+                    except PlaylistError:
                         LOG.error('Could not modify playqueue positions')
                         LOG.error('This is likely caused by mixing audio and '
                                   'video tracks in the Kodi playqueue')
                     del old[j], index[j]
                     break
             else:
+                playlistitem = PlaylistItem(kodi_item=new_item)
                 LOG.debug('Detected new Kodi element at position %s: %s ',
-                          i, new_item)
+                          i, playlistitem)
                 try:
                     if playqueue.id is None:
-                        PL.init_plex_playqueue(playqueue, kodi_item=new_item)
+                        playqueue.init(playlistitem)
                     else:
-                        PL.add_item_to_plex_playqueue(playqueue,
-                                                      i,
-                                                      kodi_item=new_item)
-                except PL.PlaylistError:
-                    # Could not add the element
-                    pass
+                        playqueue.plex_add_item(playlistitem, i)
+                except PlaylistError:
+                    LOG.warn('Couldnt add new item to Plex: %s', playlistitem)
                 except IndexError:
                     # This is really a hack - happens when using Addon Paths
                     # and repeatedly  starting the same element. Kodi will then
@@ -206,8 +102,8 @@ class PlayqueueMonitor(backgroundthread.KillableThread):
                 return
             LOG.debug('Detected deletion of playqueue element at pos %s', i)
             try:
-                PL.delete_playlist_item_from_PMS(playqueue, i)
-            except PL.PlaylistError:
+                playqueue.plex_remove_item(i)
+            except PlaylistError:
                 LOG.error('Could not delete PMS element from position %s', i)
                 LOG.error('This is likely caused by mixing audio and '
                           'video tracks in the Kodi playqueue')
